Stop using launch templates due to bug
Terraform always reports the launch template user data has changed even when it hasn't. There are multiple open long standing bugs for this. The only fix is to not use launch templates and instead set it on the ec2 instance itself. Unfortunately this means it will not attempt to replace the instance if the user data changes.