Add eu-west-1 as a supported region (#83)
### Context
Lambdas can only be run from a bucket in the target region. Meaning you
can't deploy a lambda in us-east-1 when the zip for the source code is
in a bucket in us-west-2.

### Description
Added support for eu-west-1 in the variable validation. Note, this will
only work when our next release goes out.