Fixes for Brainstore apt timeout and ephemeral volume detection (#82)
### Context
Apt:
Sometimes when brainstore instances start up they fail to install
certain things due to another process already holding the apt lock. The
default apt behavior is to instantly fail and continue on with
additional steps in the user-data init script. This would manifest as
Brainstore instances that don't have certain things installed. Unsure
what is running in the background that is using apt. Likely something
like Ubuntu auto-upgrades.

Ephemeral Volume:
We were mounting the ephemeral drive by assuming that /dev/nvme0n1 was
always the ephemeral volume. That is not the case and even between
reboots it can end up being at /dev/nvme1n1 instead.

### Description
There is a hidden apt config that allows apt to wait to acquire the
lock. It isn't even documented but it has existing since Ubuntu 20.04.
This PR adds it to the global apt.conf so that all commands use it by
default. It should hopefully help reduce the chance of this happening in
the future.

User-data script is now using the nvme cli to detect which is the
instance volume and then formatting and mounting it by UUID rather than
device name. The boot process will also fail early rather than
continuing.