Add missing env vars to AI Proxy (#87)
The API Handler and AI Proxy were not sharing common env vars and
required manually updating both of them. We missed adding some important
Brainstore env vars to the AI Proxy. This PR changes it so there is now
a common shared set of env vars. This fixes a current issue with the AI
Proxy not having access to the Brainstore Realtime WAL bucket and also
helps prevent future mistakes like this.

Additional Changes:
* Added the Brainstore bucket to outputs